Move CC functions to Egraph methods
Convert make_root, proof_forest_backtrack, leastcommonancestor,
cc_union (renamed from union), union_predecessors, and
union_process_assignment from free functions in congruence_closure.rs
into pub(crate) methods on Egraph.

congruence_closure.rs now only contains process_assignment,
find_if_eq_diseq, add_parent, get_parent, and get_child — these are
either solver-level (process_assignment) or pure helpers on
ProofForestEdge.

union_process_assignment still calls process_assignment (with a TODO
to replace with watch-based propagation).
