+++ src/egraphs/congruence_closure.rs
@@ -6,16 +6,16 @@
 use crate::datatypes::axioms::{learn_ctor_selector_clauses, learn_or_not_term_tester_term};
 use crate::egraphs::proofforest::ProofForestEdge;
 use crate::utils::{
-    DeterministicHashMap, DeterministicHashSet, FastDeterministicHashMap, fmt_termlist,
+    DeterministicHashMap, DeterministicHashSet,
 };
 use yaspar_ir::ast::alg::CheckIdentifier;
 use yaspar_ir::ast::{
-    FetchSort, HasArena, IdentifierKind, Monomorphization, Repr, Term, TermAllocator,
+    FetchSort, HasArena, IdentifierKind, Monomorphization, Repr, Term,
 };
 
 use crate::debug_println;
 use crate::egraphs::datastructures::{
-    Assertion, CanonicalForm, ConstructorType::*, DisequalTerm, Predecessor,
+    Assertion, ConstructorType::*,
 };
 use crate::egraphs::egraph::{Egraph, valid_hash};
 use crate::egraphs::unionfind::ProofTracker;
@@ -65,10 +65,9 @@ pub fn process_assignment(
             t,
             egraph.true_term
         );
-        if let Some(negated_model) = union(
+        if let Some(negated_model) = egraph.cc_union(
             *t,
             egraph.true_term,
-            egraph,
             res,
             level,
             fixed,
@@ -101,10 +100,9 @@ pub fn process_assignment(
             t,
             egraph.false_term
         );
-        if let Some(negated_model) = union(
+        if let Some(negated_model) = egraph.cc_union(
             *t,
             egraph.false_term,
-            egraph,
             res,
             level,
             fixed,
@@ -214,7 +212,7 @@ pub fn process_assignment(
                     children: DeterministicHashSet::new(),
                 }
             };
-            union(t1, t2, egraph, reason, level, fixed, from_quantifier)
+            egraph.cc_union(t1, t2, reason, level, fixed, from_quantifier)
         }
         Assertion::Disequality {
             t1,
@@ -234,7 +232,7 @@ pub fn process_assignment(
             debug_println!(10, 0, "{}", egraph);
 
             if let Some(negated_model) =
-                leastcommonancestor(t1, t2, egraph, &mut ProofTracker::new())
+                egraph.leastcommonancestor(t1, t2, &mut ProofTracker::new())
             {
                 let mut model_terms: Vec<i32> = negated_model
                     .into_iter()
@@ -270,7 +268,7 @@ pub fn process_assignment(
                         hash
                     );
                     if let Some(negated_model) =
-                        leastcommonancestor(t1, t2, egraph, &mut ProofTracker::new())
+                        egraph.leastcommonancestor(t1, t2, &mut ProofTracker::new())
                     {
                         let mut model_terms: Vec<i32> = negated_model
                             .into_iter()
@@ -309,7 +307,7 @@ pub fn process_assignment(
     //  debug_println!(4, 0, "{}", egraph);
     debug_println!(10, 0, "Checking if true = false {}", egraph);
     if let Some(negated_model) =
-        leastcommonancestor(egraph.true_term, egraph.false_term, egraph, &mut tracker)
+        egraph.leastcommonancestor(egraph.true_term, egraph.false_term, &mut tracker)
     {
         let negated_model_terms: Vec<i32> = negated_model
             .into_iter()
@@ -459,173 +457,6 @@ pub fn find_if_eq_diseq<'a>(
     }
 }
 
-fn leastcommonancestor_helper(
-    u: u64,
-    v: u64,
-    egraph: &Egraph,
-    tracker: &mut ProofTracker,
-    indent: usize,
-) -> Option<Vec<(u64, u64)>> {
-    debug_println!(
-        20,
-        indent,
-        "checking the equality of {} and {}",
-        egraph.get_term(u),
-        egraph.get_term(v)
-    );
-    let mut visited = DeterministicHashSet::default();
-
-    // Follow and record the path from u to its root
-    let mut path_from_u = vec![];
-    let mut curr = u;
-
-    let max_recursion_depth = 100;
-    if indent > max_recursion_depth {
-        debug_println!(11, 0, "We have the proof forest :{}", egraph);
-        panic!("Should not have this many recusive calls to LCH");
-    }
-    loop {
-        // debug_println!(16, 0, "curr {}", egraph.get_term(curr));
-        let parent = egraph.proof_forest[curr as usize].clone();
-        visited.insert(curr);
-        if let ProofForestEdge::Root {
-            size: _,
-            child: _,
-            disequalities: _,
-            children: _,
-        } = parent
-        {
-            visited.insert(curr);
-            break;
-        }
-        curr = get_parent(&parent);
-        path_from_u.push(parent);
-    }
-
-    // Follow path from v until we hit a node visited by u's path
-    let mut path_from_v = vec![];
-    curr = v;
-    let mut parent: ProofForestEdge;
-    loop {
-        // debug_println!(16, 0, "curr [round 2] {}: {}", egraph.get_term(curr), curr);
-        parent = egraph.proof_forest[curr as usize].clone();
-        if visited.contains(&curr) {
-            break;
-        }
-        if let ProofForestEdge::Root {
-            size: _,
-            child: _,
-            disequalities: _,
-            children: _,
-        } = parent
-        {
-            // if this happens, they are not in the same equivalence class and thus unequal
-            return None;
-        }
-        curr = get_parent(&parent);
-        path_from_v.push(parent);
-    }
-
-    let mut proof: Vec<ProofForestEdge> = Vec::new();
-    proof.extend(
-        path_from_u
-            .iter()
-            .take_while(|x| **x != parent)
-            .cloned()
-            .collect::<Vec<ProofForestEdge>>(),
-    );
-    proof.extend(path_from_v);
-
-    assert!(visited.contains(&curr));
-
-    let mut final_proof = vec![];
-    let mut proof_congruences = vec![];
-
-    debug_println!(11, indent + 1, "We get the unprocessed proof {:?}", proof);
-    debug_println!(16, indent + 1, "We have the proof:");
-    // For each pair in the proof path
-    for proof_term in proof {
-        match proof_term {
-            ProofForestEdge::Root {
-                size: _,
-                child: _,
-                disequalities: _,
-                children: _,
-            } => {
-                eprintln!("ERROR: Root should not be processed");
-                std::process::exit(1);
-            }
-            ProofForestEdge::Congruence { pairs, .. } => {
-                if is_important(20) {
-                    debug_println!(20, indent + 12, "Congruence ");
-                    for (t1, t2) in pairs.clone() {
-                        debug_println!(
-                            20,
-                            indent + 12,
-                            "{} [{}] ~ {} [{}] ",
-                            egraph.get_term(t1),
-                            t1,
-                            egraph.get_term(t2),
-                            t2
-                        );
-                    }
-                }
-                proof_congruences.push(pairs);
-            }
-            ProofForestEdge::Equality { term, .. } => {
-                if let Some((t1, t2)) = term {
-                    // if the term is None, then it comes from setting a term = annotated term, and we do not need it for conflict
-                    debug_println!(
-                        20,
-                        indent + 12,
-                        "Equality {} [{}] = {} [{}]",
-                        egraph.get_term(t1),
-                        t1,
-                        egraph.get_term(t2),
-                        t2
-                    );
-                    if tracker.union(t1, t2) {
-                        final_proof.push((t1, t2));
-                    }
-                    debug_println!(
-                        11,
-                        1,
-                        "We have the current final proof is: {:?}",
-                        final_proof
-                    )
-                }
-            }
-        }
-    }
-
-    // need to handle equalities first then congurences
-    for pairs in proof_congruences {
-        for pair in pairs {
-            if let Some(subproof) =
-                leastcommonancestor_helper(pair.0, pair.1, egraph, tracker, indent + 1)
-            {
-                final_proof.extend(subproof);
-            }
-        }
-    }
-    Some(final_proof)
-}
-
-pub fn leastcommonancestor(
-    u: u64,
-    v: u64,
-    egraph: &Egraph,
-    tracker: &mut ProofTracker,
-) -> Option<Vec<(u64, u64)>> {
-    debug_println!(
-        11,
-        1,
-        "Finding least common ancestor for {} and {}",
-        egraph.get_term(u),
-        egraph.get_term(v)
-    );
-    leastcommonancestor_helper(u, v, egraph, tracker, 0)
-}
 
 pub fn add_parent(
     proof_parent: ProofForestEdge,
@@ -707,777 +538,4 @@ pub fn get_child(proof_parent: &ProofForestEdge) -> u64 {
     }
 }
 
-/// UNION operation for union-find
-///
-/// As an optimization, could potentially add the tree
-/// with smaller depth as a child of tree with larget depth.
-/// This will result in the tree depth being bound by log(n)
-///
-/// Doing the union thing where we combine trees by making y the root
-/// and adding edge x -> y. Good for recovering proof at the end,
-/// but this could double/triple the max tree size at each iteration
-///
-/// design decision: don't have eager updates for equivalence class and inverting tree
-pub fn union(
-    x: u64,
-    y: u64,
-    egraph: &mut Egraph,
-    proof_parent: ProofForestEdge,
-    level: usize,
-    fixed: bool,
-    from_quantifier: bool,
-) -> Option<Vec<Vec<i32>>> {
-    let x_root = egraph.find(x);
-    let y_root = egraph.find(y);
-    debug_println!(6, 1, "{}", egraph);
-    debug_println!(6, 0, "before1");
-    debug_println!(
-        22,
-        1,
-        "Unioning vertices [{}] {}  and [{}] {}  (roots: {} [{}] and {} [{}]) at level {} with {}",
-        x,
-        egraph.get_term(x),
-        y,
-        egraph.get_term(y),
-        x_root,
-        egraph.get_term(x_root),
-        y_root,
-        egraph.get_term(y_root),
-        level,
-        proof_parent
-    );
-    debug_println!(11, 0, "{}", egraph);
-    assert_eq!(
-        egraph.get_term(x).get_sort(egraph),
-        egraph.get_term(y).get_sort(egraph),
-        "We are comparing terms {} and {}",
-        egraph.get_term(x),
-        egraph.get_term(y)
-    );
 
-    if x_root == y_root {
-        debug_println!(
-            16,
-            2,
-            "{} and {} are already in the same equivalence class",
-            egraph.get_term(x),
-            egraph.get_term(y)
-        );
-        return None;
-    }
-
-    // keep track of original proof_parent
-    let _proof_parent_original = proof_parent.clone();
-
-    // making x the parent of y ~> could also do this based on relative depth of x and y tree
-    let proof_parent: ProofForestEdge =
-        add_parent(proof_parent, x, y, level, egraph.predecessor_hash);
-
-    let y_root_parent = &egraph.proof_forest[y_root as usize];
-
-    if !fixed {
-        // not adding fixed levels to backtracking based on what Armin said
-        debug_println!(
-            16,
-            0,
-            "BACKTTRACK STACK: adding equalitity between {} and {} with y_root: {} at level {}",
-            egraph.get_term(x),
-            egraph.get_term(y),
-            egraph.get_term(y_root),
-            level
-        );
-        egraph.proof_forest_backtrack_stack.push((
-            level,
-            proof_parent.clone(),
-            y_root,
-            y_root_parent.clone(),
-        ));
-    }
-
-    debug_println!(
-        16,
-        2,
-        "Making {} the root of its equivalence class [previously was {}]",
-        egraph.get_term(y),
-        egraph.get_term(y_root)
-    );
-    make_root(y, proof_parent, egraph);
-
-    // need to add the new disequalities into x_root
-    // TODO: could also clean up some backtracking stuff here, probably want to factor this into its own function
-    let (x_root_disequalities_edge, y_root_disequalities_edge) = if x_root > y_root {
-        let split = egraph.proof_forest.split_at_mut(x_root as usize);
-        (&mut split.1[0], &split.0[y_root as usize])
-    } else {
-        let split = egraph.proof_forest.split_at_mut(y_root as usize);
-        (&mut split.0[x_root as usize], &split.1[0])
-    };
-
-    let y_root_disequalities = y_root_disequalities_edge.disequalities();
-    let x_root_disequalities = x_root_disequalities_edge.disequalities_mut();
-
-    // when we copy things over, make sure we only copy things over that are valid and that we are updating the hash/level -> this caused some very tricky bugs
-    // TODO: write helper functions to make copying over hased things easier
-    for (key, value) in y_root_disequalities {
-        // make sure we update the disequality level
-        if value.hash >= egraph.predecessor_level[value.level]
-            || value.hash == 0
-            || value.level == 0
-        {
-            // added value.level == 0 since I think all hashes should be valid at level 0
-            // TODO: borrowing issue so I can't use valid_hash function
-
-            // we can have that we introduce a new equality via eclass option after a quantifier instantiation
-            // this equality could be at level 0
-            // but then it's possible that there are disequalities that get copied over such that one of the disequalities are at a level higher than 0
-            let (diseq_level, diseq_hash) = if value.level > level {
-                (value.level, value.hash)
-            } else {
-                (level, egraph.predecessor_level[level])
-            };
-
-            let new_value = DisequalTerm {
-                term: value.term,
-                diseq_lit: value.diseq_lit,
-                level: diseq_level,
-                hash: diseq_hash,
-                original_disequality: value.original_disequality,
-            };
-            // this assert is obviously not true as x_root_disequalities could contain key
-            // but then why is it not a problem that we are overwriting it
-            // assert!(!x_root_disequalities.contains_key(key));
-
-            if let Some(x_disequality) = x_root_disequalities.get(key)
-                && (x_disequality.hash >= egraph.predecessor_level[x_disequality.level]
-                    || x_disequality.hash == 0)
-            {
-                debug_println!(
-                    12,
-                    0,
-                    "Skipping disequality {} : {} to {} at level {}",
-                    key,
-                    new_value,
-                    x_root,
-                    level
-                );
-                continue;
-            }
-
-            debug_println!(
-                12,
-                0,
-                "Adding disequality {} : {} to {} at level {}",
-                key,
-                new_value,
-                x_root,
-                level
-            );
-            x_root_disequalities.insert(*key, new_value);
-        }
-    }
-
-    // basically checking if the current equality that we just added violated any earlier disequalities and if it did, we learn a conflict clause
-    // TODO: now I am actually not sure if disequality checking is really necessary
-    // kind've a weird way to do it since we have already unioned, we are just checking if two things are unequal to themselves
-    debug_println!(
-        5,
-        0,
-        "A. Checking the equality {} = {} with disequalities {:?}",
-        egraph.get_term(x),
-        egraph.get_term(y),
-        egraph.proof_forest[x as usize].disequalities()
-    );
-    if let Some(disequality) = egraph.check_self_disequality(x_root).clone() {
-        debug_println!(
-            11,
-            0,
-            "B. Checking the equality {} = {} with disequality {} != {}",
-            egraph.get_term(x),
-            egraph.get_term(y),
-            egraph.get_term(disequality.original_disequality.0),
-            egraph.get_term(disequality.original_disequality.1)
-        );
-        let mut tracker = ProofTracker::new();
-        // tracker.initialize_tracker(egraph.num_vars as u64);
-        if let Some(negated_model) = leastcommonancestor(
-            disequality.original_disequality.0,
-            disequality.original_disequality.1,
-            egraph,
-            &mut tracker,
-        ) {
-            let mut model_terms: Vec<i32> = negated_model
-                .into_iter()
-                .map(|x| -egraph.make_eq(x.0, x.1))
-                .collect();
-            model_terms.push(-disequality.diseq_lit);
-            debug_println!(
-                7,
-                1,
-                "Contradiction found [3]: {:?} [{:?}]",
-                model_terms
-                    .iter()
-                    .map(|x| egraph.get_term_from_lit(*x))
-                    .collect::<Vec<_>>(),
-                model_terms
-            );
-            return Some(vec![model_terms]); // Return negated model as the contradiction explanation (todo: ideally would want to have a way to specify this is a model to the thing we pass to, but this is not possible at least with our current setup)
-        } else {
-            debug_println!(16, 0, "{}", egraph);
-            panic!(
-                "Should have found a equality between {} [root: {}] and {} [root: {}]",
-                egraph.get_term(disequality.original_disequality.0),
-                egraph.get_term(egraph.find(disequality.original_disequality.0)),
-                egraph.get_term(disequality.original_disequality.1),
-                egraph.get_term(egraph.find(disequality.original_disequality.1)),
-            );
-        }
-    }
-
-    union_predecessors(x_root, y_root, egraph, level, fixed, from_quantifier)
-}
-
-/// Make vertex the root of its tree
-fn make_root(vertex: u64, proof_parent: ProofForestEdge, egraph: &mut Egraph) {
-    debug_println!(
-        16,
-        0,
-        "Making {} the root with proof_parent {}",
-        egraph.get_term(vertex),
-        proof_parent
-    );
-    let old_parent = egraph.proof_forest[vertex as usize].clone();
-    let disequalities = match old_parent {
-        ProofForestEdge::Root { disequalities, .. } => disequalities,
-        ProofForestEdge::Congruence {
-            size: _, // note we don't use size right now, but could in the future
-            pairs,
-            parent,
-            child,
-            disequalities,
-            level,
-            hash,
-            children,
-        } => {
-            assert_eq!(child, vertex);
-            make_root(
-                parent,
-                ProofForestEdge::Congruence {
-                    size: 0,
-                    pairs,
-                    parent: vertex,
-                    child: parent,
-                    disequalities: DeterministicHashMap::new(),
-                    level,
-                    hash,
-                    children,
-                },
-                egraph,
-            );
-            disequalities
-        }
-        ProofForestEdge::Equality {
-            term,
-            parent,
-            child,
-            disequalities,
-            level,
-            hash,
-            children,
-            ..
-        } => {
-            assert_eq!(child, vertex);
-            make_root(
-                parent,
-                ProofForestEdge::Equality {
-                    size: 0,
-                    term,
-                    parent: vertex,
-                    child: parent,
-                    disequalities: DeterministicHashMap::new(),
-                    level,
-                    hash,
-                    children,
-                },
-                egraph,
-            );
-            disequalities
-        }
-    };
-    let new_proof_parent = proof_parent.set_disequalities(disequalities);
-    egraph.proof_forest[vertex as usize] = new_proof_parent;
-}
-
-/// Given u and v (roots of u_original and v_original), check the predecessors of
-/// each of these and union them if they have become equal
-///
-/// TODO: I probably actually don't want this to delete all of the predecessors of u because it will screw up backtracking
-/// you only have to do it for predecessor terms that are roots of a congruent class
-/// once you merge two predecessor states, then you don't need to look at it until you backtrack
-///
-/// TODO: need to implement a backtracking where I change the predecessor hash
-fn union_predecessors(
-    u: u64,
-    v: u64,
-    egraph: &mut Egraph,
-    level: usize,
-    fixed: bool,
-    from_quantifier: bool,
-) -> Option<Vec<Vec<i32>>> {
-    debug_println!(
-        11,
-        1,
-        "Unioning predecessors of {} [{}, Predecessors: {}] and {} [{}, Predecessors: {}]",
-        egraph.get_term(u),
-        u,
-        fmt_termlist(
-            egraph.predecessors[u as usize]
-                .keys()
-                .map(|x| egraph.get_term(*x))
-                .collect::<Vec<_>>()
-        ),
-        egraph.get_term(v),
-        v,
-        fmt_termlist(
-            egraph.predecessors[v as usize]
-                .keys()
-                .map(|x| egraph.get_term(*x))
-                .collect::<Vec<_>>()
-        )
-    );
-
-    debug_assert!(u != v);
-    debug_assert!(egraph.find(u) == u);
-
-    // Move u's and v's predecessor maps out of the egraph so we can iterate
-    // without cloning. Both slots are restored before any re-entrant call
-    // (add_predecessor / union_process_assignment) can observe them.
-    let mut predecessors_u = std::mem::take(&mut egraph.predecessors[u as usize]);
-    let predecessors_v = std::mem::take(&mut egraph.predecessors[v as usize]);
-
-    let mut canonical_forms_u: FastDeterministicHashMap<_, Vec<(Vec<u64>, u64)>> =
-        FastDeterministicHashMap::default();
-
-    // Stale entries are dropped in-place via retain before iterating.
-    predecessors_u.retain(|_, p| {
-        let keep = valid_hash(p.hash, p.level, &egraph.predecessor_level);
-        if !keep {
-            debug_println!(
-                11,
-                2,
-                "CANONICAL_FORMS_U: Skipping predecessor {} of {} [original: {}] as it has hash {} at level {} and correct hash is {}",
-                egraph.get_term(p.predecessor),
-                egraph.get_term(u),
-                egraph.get_term(p.inner_term),
-                p.hash,
-                p.level,
-                egraph.predecessor_level[p.level]
-            );
-        }
-        keep
-    });
-
-    for (_pred_u_key, predecessor_u) in predecessors_u.iter() {
-        debug_println!(
-            11,
-            2,
-            "1.We are in union_predecessors trying to get term for {}",
-            egraph.get_term(predecessor_u.predecessor)
-        );
-
-        // checking if the ite leads to a contradiction
-        // if let Some(negated_model) =
-        //     union_process_ite(&egraph.get_term(predecessor_u.predecessor), egraph, level, from_quantifier)
-        // {
-        //      debug_println!(
-        //         4,
-        //         3,
-        //         "M. Contradiction found in union_predecessors, we have the following negated_model: {:?}",
-        //         negated_model
-        //     );
-        //     return Some(negated_model);
-        // }
-
-        if let Some(CanonicalForm {
-            original_subterms,
-            op,
-            canonical_subterms,
-        }) = egraph.get_canonical_form(predecessor_u.predecessor, level)
-        {
-            let canonical_form = (op, canonical_subterms);
-            debug_println!(
-                11,
-                4,
-                "We are adding in the canonical_form {:?}",
-                canonical_form
-            );
-            if let Some(forms) = canonical_forms_u.get_mut(&canonical_form) {
-                forms.push((original_subterms, predecessor_u.predecessor))
-            } else {
-                canonical_forms_u.insert(
-                    canonical_form,
-                    vec![(original_subterms, predecessor_u.predecessor)],
-                );
-            }
-        }
-    }
-
-    debug_println!(
-        11,
-        4,
-        "2.We have the canonical_forms_u {:?}",
-        canonical_forms_u
-    );
-
-    // Restore u's slot before calling add_predecessor below, which writes to it.
-    egraph.predecessors[u as usize] = predecessors_u;
-
-    // basically the issue was that in `union_predecessors` when you create a `canonical_term_u`,
-    // you fix it, but then you compare to a for loop iterating through all terms in v and iteratively
-    // computing the canonical_term_v, but this could change as you are iterating through the loop
-    // so we want to precompute the canonical terms of v.
-    //
-    // Precompute: store (key, predecessor_id, canonical_form) per entry.
-    // No Predecessor clone — just the scalar `predecessor` field needed downstream.
-    let mut predecessor_v_canonical_forms: Vec<(u64, u64, Option<CanonicalForm>)> =
-        Vec::with_capacity(predecessors_v.len());
-    for (pred_v_key, predecessor_v) in predecessors_v.iter() {
-        let canonical_form = egraph.get_canonical_form(predecessor_v.predecessor, level);
-        predecessor_v_canonical_forms.push((
-            *pred_v_key,
-            predecessor_v.predecessor,
-            canonical_form,
-        ));
-    }
-
-    // moving predecessors from v to u
-    for (pred_key, pred_val) in predecessors_v.iter() {
-        debug_println!(
-            11,
-            0,
-            "We are are adding predecessor {} (of  {}) to {} [level: {}, hash: {}]",
-            egraph.get_term(*pred_key),
-            egraph.get_term(pred_val.inner_term),
-            egraph.get_term(u),
-            level,
-            egraph.predecessor_hash
-        );
-        let new_pred = Predecessor {
-            level,
-            hash: egraph.predecessor_hash,
-            predecessor: *pred_key,
-            inner_term: pred_val.inner_term,
-        };
-        egraph.add_predecessor(u, *pred_key, new_pred);
-    }
-
-    // Restore v before the consuming loop — union_process_assignment can
-    // re-enter and read/write egraph.predecessors[v].
-    egraph.predecessors[v as usize] = predecessors_v;
-
-    for (pred_v_key, pred_predecessor, canonical_form_v) in predecessor_v_canonical_forms {
-        // Look up the predecessor's validity from the restored v slot.
-        // Extract only scalar fields — no Predecessor clone.
-        let (pred_hash, pred_level, pred_inner_term) =
-            match egraph.predecessors[v as usize].get(&pred_v_key) {
-                Some(p) => (p.hash, p.level, p.inner_term),
-                None => continue, // removed by a prior iteration
-            };
-        if !valid_hash(pred_hash, pred_level, &egraph.predecessor_level) {
-            debug_println!(
-                11,
-                5,
-                "Skipping predecessor {} of {} [original: {}] as it has hash {} at level {} and correct hash is {}",
-                egraph.get_term(pred_predecessor),
-                egraph.get_term(v),
-                egraph.get_term(pred_inner_term),
-                pred_hash,
-                pred_level,
-                egraph.predecessor_level[pred_level]
-            );
-            debug_println!(
-                11,
-                5,
-                "The current level is {} and hash is {}",
-                level,
-                egraph.predecessor_hash
-            );
-            egraph.predecessors[v as usize].remove(&pred_v_key);
-            continue;
-        }
-        debug_println!(
-            11,
-            3,
-            "L. We are in union_predecessors trying to get term for {}",
-            egraph.get_term(pred_predecessor)
-        );
-
-        if let Some(CanonicalForm {
-            original_subterms,
-            op,
-            canonical_subterms,
-        }) = canonical_form_v
-        {
-            let canonical_form = (op, canonical_subterms);
-            debug_println!(
-                11,
-                6,
-                "3. We are in union_predecessors for v and have canonical form {:?} for {}",
-                canonical_form,
-                egraph.get_term(pred_predecessor)
-            );
-            if let Some(u_forms) = canonical_forms_u.get(&canonical_form) {
-                debug_println!(5, 0, "We have the following u_forms {:?}", u_forms);
-                for (u_original_subterms, canonical_form_u) in u_forms {
-                    debug_println!(
-                        16,
-                        0,
-                        "We are actually merging the two predecessors {} and {}",
-                        egraph.get_term(*canonical_form_u),
-                        egraph.get_term(pred_predecessor)
-                    );
-                    if is_important(16) {
-                        debug_println!(16, 0, "We have u_original_subterms: ");
-                        for term in u_original_subterms {
-                            debug_println!(16, 4, "{}", egraph.get_term(*term));
-                        }
-                        debug_println!(16, 0, "We have original_subterms: ");
-                        for term in original_subterms.clone() {
-                            debug_println!(16, 4, "{}", egraph.get_term(term));
-                        }
-                    }
-
-                    let terms_pairwise = u_original_subterms
-                        .clone()
-                        .into_iter()
-                        .zip(original_subterms.clone())
-                        .collect::<Vec<(u64, u64)>>();
-                    let proof_parent = ProofForestEdge::Congruence {
-                        size: 0,
-                        pairs: terms_pairwise,
-                        parent: 0,
-                        child: 0,
-                        disequalities: DeterministicHashMap::new(),
-                        level,
-                        hash: egraph.predecessor_hash,
-                        children: DeterministicHashSet::new(),
-                    }; // TODO: I can't have a child of -1 anymore, but I think doing it like this is correct
-
-                    // TODO: this is a hackish way to add this if the equality occurs in the formula
-                    if let Some(negated_model) = union_process_assignment(
-                        // u_original,
-                        // v_original,
-                        *canonical_form_u,
-                        pred_predecessor,
-                        egraph,
-                        proof_parent,
-                        level,
-                        fixed,
-                        from_quantifier,
-                    ) {
-                        debug_println!(
-                            11,
-                            5,
-                            "[exiting union_pred] of {} and {} In union_predecessors, we have the following negated_model: {:?}",
-                            egraph.get_term(u),
-                            egraph.get_term(v),
-                            negated_model
-                        );
-                        return Some(negated_model);
-                    }
-                }
-            }
-        }
-    }
-    debug_println!(
-        11,
-        0,
-        "[exiting union_pred] of {} and {} with None",
-        egraph.get_term(u),
-        egraph.get_term(v)
-    );
-    None
-}
-
-/// Helper function called by union_predecessors that represents a recursive call
-/// to union
-///
-/// TODO: would be cleaner to have union_predecessors just call union
-fn union_process_assignment(
-    x: u64,
-    y: u64,
-    egraph: &mut Egraph,
-    proof_parent: ProofForestEdge,
-    level: usize,
-    fixed: bool,
-    from_quantifier: bool,
-) -> Option<Vec<Vec<i32>>> {
-    debug_println!(6, 0, "before4");
-    let new_assignment = egraph.eq(egraph.get_term(x), egraph.get_term(y));
-    // if there is a new assignment, we need to check if the equality term exists, if it does we need to work on that
-    // otherwise we can just consider the union of these two terms
-    if let Some(new_assignment_lit) = egraph.cnf_cache.var_map.get(&new_assignment.uid()) {
-        // note we don't want reason to be the above thing because the explanation is still the same as teh explanation before
-        let reason = proof_parent;
-        debug_println!(
-            16,
-            0,
-            "We are in union_process_assignment trying to process assignment for x: {} [{}] and y: {} [{}] and fixed {}",
-            egraph.get_term(x),
-            x,
-            egraph.get_term(y),
-            y,
-            fixed
-        );
-        // if let ProofForestEdge::Congruence{pairs, ..} = proof_parent && pairs.len() == 1 {
-        //     debug_println!(
-        //     16,
-        //     0,
-        //     "We have x_original: {} [{}] and y_original {} [{}] and Congruence with {} [{}] and {} [{}]",
-        //     egraph.get_term(x_original),
-        //     x_original,
-        //     egraph.get_term(y_original),
-        //     y_original,
-        //     egraph.get_term(pairs[0].0),
-        //     pairs[0].0,
-        //     egraph.get_term(pairs[0].1),
-        //     pairs[0].1
-        //     )
-        // }
-        let negated_model_additional_constraints_opt = process_assignment(
-            *new_assignment_lit,
-            egraph,
-            level,
-            false,
-            from_quantifier,
-            Some(reason),
-        );
-        // assert!(additional_constraints_opt.is_none()); // this should be done becaue right now we only get new constraints for a datatype literal
-        if let Some(negated_model) = negated_model_additional_constraints_opt {
-            debug_println!(
-                6,
-                0,
-                "We have the following negated_model: {:?}",
-                negated_model
-            );
-            return Some(negated_model);
-        };
-    } else {
-        debug_println!(
-            16,
-            0,
-            "We are in union_process_assignment trying to union {} and {} with fixed {}",
-            egraph.get_term(x),
-            egraph.get_term(y),
-            fixed
-        );
-        if let Some(negated_model) =
-            union(x, y, egraph, proof_parent, level, fixed, from_quantifier)
-        {
-            return Some(negated_model);
-        }
-    };
-    None
-}
-
-pub fn proof_forest_backtrack(
-    equality: ProofForestEdge,
-    y: u64,
-    y_parent: ProofForestEdge,
-    egraph: &mut Egraph,
-) {
-    let child = &get_child(&equality);
-    let child_edge = egraph.proof_forest[*child as usize].clone();
-    let parent = &get_parent(&equality);
-    let parent_edge = egraph.proof_forest[*parent as usize].clone();
-
-    assert_eq!(egraph.find(*child), egraph.find(*parent));
-
-    debug_println!(
-        16,
-        0,
-        "Backtracking on {} with child {} and parent {} and y_term {}",
-        equality,
-        egraph.get_term(*child),
-        egraph.get_term(*parent),
-        egraph.get_term(y)
-    );
-
-    debug_println!(
-        6,
-        0,
-        "We are in proof_forest_backtrack trying to get term for {:?}",
-        child
-    );
-
-    debug_println!(
-        6,
-        0,
-        "We have child_edge {:?}, parent_edge {:?} and equality {:?}",
-        child_edge,
-        parent_edge,
-        equality
-    );
-    let (child, child_edge, _parent, _parent_edge) = if child_edge != equality {
-        // if this is the case, the edge has been reversed at some point
-        // and the child is actually the parent
-        debug_println!(6, 0, "we are reversing the edge");
-        debug_println!(10, 0, "{}", egraph);
-        assert_eq!(get_parent(&parent_edge), get_child(&equality));
-        debug_println!(6, 0, "after first assert");
-        assert_eq!(get_child(&parent_edge), get_parent(&equality));
-        (parent, parent_edge, child, child_edge)
-    } else {
-        (child, child_edge, parent, parent_edge)
-    };
-
-    debug_println!(
-        6,
-        0,
-        "We are setting the predecessors of the child {} to {:?}",
-        egraph.get_term(*child),
-        egraph.predecessors[*child as usize]
-    );
-
-    let childs_child = get_child(&child_edge);
-
-    // we are adding disequalities from the "parent" edge to the child
-    let mut new_disequalities = DeterministicHashMap::new();
-    for (k, v) in child_edge.disequalities().iter() {
-        if valid_hash(v.hash, v.level, &egraph.predecessor_level) {
-            debug_println!(11, 0, "Keeping disequality {}: {} in {}", k, v, child);
-            new_disequalities.insert(*k, v.clone());
-        } else {
-            debug_println!(
-                11,
-                0,
-                "Removing disequality {}: {} from {}",
-                k,
-                v,
-                child_edge
-            );
-        }
-    }
-
-    // do the same for predecessors
-    let child_root = ProofForestEdge::Root {
-        size: 0,
-        child: childs_child,
-        disequalities: new_disequalities,
-        children: DeterministicHashSet::new(), // todo: need to actually update children
-    };
-
-    egraph.proof_forest[*child as usize] = child_root;
-
-    // when backtracking undo the make_root from union
-    debug_println!(
-        16,
-        0,
-        "Making {} the root on a backtrack",
-        egraph.get_term(y)
-    );
-    make_root(y, y_parent, egraph);
-
-    // for (pred_key, pred_value) in _ {
-    //     egraph.add_predecessor(y, new_pred_key, new_pred);
-    // }
-}
